+++ config/initializers/assets.rb
@@ -12,5 +12,4 @@ Rails.application.config.assets.version = "1.0"
 # Rails.application.config.assets.precompile += %w( admin.js admin.css )
 Rails.application.config.assets.precompile += %w(bootstrap.min.js popper.js)
 Rails.application.config.assets.precompile << "chart.js/dist/chart.umd.js"
-Rails.application.config.assets.precompile << "@popperjs/core/dist/umd/popper.min.js"
 Rails.application.config.assets.paths << Rails.root.join("node_modules")